Big code clean up/refactor + package as DLL
