--- PetaPoco/App.config
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-	<connectionStrings>
-		<add name="mysql" connectionString="server=jab;database=jab;user id=jab;password=jab;Allow User Variables=true" providerName="MySql.Data.MySqlClient"/>
-		<add name="sqlserver" connectionString="Data Source=MACBOOK\SQLEXPRESS;Initial Catalog=testdb;Integrated Security=True" providerName="System.Data.SqlClient"/>
-		<add name="sqlserverce" connectionString="Data Source=C:\Users\bradr\dev\Source\PetaPoco\PetaPoco.Tests\petapoco.sdf" providerName="System.Data.SqlServerCe.4.0"/>
-		<add name="postgresql" connectionString="Server=127.0.0.1;User id=postgres;password=password01;Database=tempdb;" providerName="Npgsql"/>
-	</connectionStrings>
-
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0"/></startup></configuration>
