+++ PetaPoco/Core/IMapper.cs
@@ -1,0 +1,54 @@
+﻿// PetaPoco - A Tiny ORMish thing for your POCO's.
+// Copyright © 2011-2012 Topten Software.  All Rights Reserved.
+ 
+using System;
+using System.Reflection;
+
+namespace PetaPoco
+{
+	// Optionally provide an implementation of this to Database.Mapper
+	public interface IMapper
+	{
+		/// <summary>
+		/// Get information about the table associated with a POCO class
+		/// </summary>
+		/// <param name="pocoType"></param>
+		/// <returns>A TableInfo instance</returns>
+		/// <remarks>
+		/// This method must return a valid TableInfo.  
+		/// To create a TableInfo from a POCO's attributes, use TableInfo.FromPoco
+		/// </remarks>
+		TableInfo GetTableInfo(Type pocoType);
+
+		/// <summary>
+		/// Get information about the column associated with a property of a POCO
+		/// </summary>
+		/// <param name="pocoProperty">The PropertyInfo of the property being queried</param>
+		/// <returns>A reference to a ColumnInfo instance, or null to ignore this property</returns>
+		/// <remarks>
+		/// To create a ColumnInfo from a property's attributes, use PropertyInfo.FromProperty
+		/// </remarks>
+		ColumnInfo GetColumnInfo(PropertyInfo pocoProperty);
+
+		/// <summary>
+		/// Supply a function to convert a database value to the correct property value
+		/// </summary>
+		/// <param name="TargetProperty">The target property</param>
+		/// <param name="SourceType">The type of data returned by the DB</param>
+		/// <returns>A Func that can do the conversion, or null for no conversion</returns>
+		Func<object, object> GetFromDbConverter(PropertyInfo TargetProperty, Type SourceType);
+
+		/// <summary>
+		/// Supply a function to convert a property value into a database value
+		/// </summary>
+		/// <param name="SourceProperty">The property to be converted</param>
+		/// <returns>A Func that can do the conversion</returns>
+		/// <remarks>
+		/// This conversion is only used for converting values from POCO's that are 
+		/// being Inserted or Updated.  
+		/// Conversion is not available for parameter values passed directly to queries.
+		/// </remarks>
+		Func<object, object> GetToDbConverter(PropertyInfo SourceProperty);
+	}
+
+}
