+++ PetaPoco/Core/MultiPocoFactory.cs
@@ -1,0 +1,170 @@
+﻿// PetaPoco - A Tiny ORMish thing for your POCO's.
+// Copyright © 2011-2012 Topten Software.  All Rights Reserved.
+ 
+using System;
+using System.Text;
+using System.Linq;
+using System.Collections.Generic;
+using System.Reflection.Emit;
+using System.Linq.Expressions;
+using System.Data;
+
+namespace PetaPoco
+{
+	class MultiPocoFactory
+	{
+		// Instance data used by the Multipoco factory delegate - essentially a list of the nested poco factories to call
+		List<Delegate> _delegates;
+		public Delegate GetItem(int index) { return _delegates[index]; }
+
+		// Automagically guess the property relationships between various POCOs and create a delegate that will set them up
+		public static object GetAutoMapper(Type[] types)
+		{
+			// Build a key
+			var kb = new StringBuilder();
+			foreach (var t in types)
+			{
+				kb.Append(t.ToString());
+				kb.Append(":");
+			}
+			var key = kb.ToString();
+
+			return AutoMappers.Get(key, () =>
+				{
+					// Create a method
+					var m = new DynamicMethod("petapoco_automapper", types[0], types, true);
+					var il = m.GetILGenerator();
+
+					for (int i = 1; i < types.Length; i++)
+					{
+						bool handled = false;
+						for (int j = i - 1; j >= 0; j--)
+						{
+							// Find the property
+							var candidates = from p in types[j].GetProperties() where p.PropertyType == types[i] select p;
+							if (candidates.Count() == 0)
+								continue;
+							if (candidates.Count() > 1)
+								throw new InvalidOperationException(string.Format("Can't auto join {0} as {1} has more than one property of type {0}", types[i], types[j]));
+
+							// Generate code
+							il.Emit(OpCodes.Ldarg_S, j);
+							il.Emit(OpCodes.Ldarg_S, i);
+							il.Emit(OpCodes.Callvirt, candidates.First().GetSetMethod(true));
+							handled = true;
+						}
+
+						if (!handled)
+							throw new InvalidOperationException(string.Format("Can't auto join {0}", types[i]));
+					}
+
+					il.Emit(OpCodes.Ldarg_0);
+					il.Emit(OpCodes.Ret);
+
+					// Cache it
+					return m.CreateDelegate(Expression.GetFuncType(types.Concat(types.Take(1)).ToArray()));
+				}
+			);
+		}
+
+		// Find the split point in a result set for two different pocos and return the poco factory for the first
+		static Delegate FindSplitPoint(Type typeThis, Type typeNext, string ConnectionString, string sql, IDataReader r, ref int pos)
+		{
+			// Last?
+			if (typeNext == null)
+				return PocoData.ForType(typeThis).GetFactory(sql, ConnectionString, pos, r.FieldCount - pos, r);
+
+			// Get PocoData for the two types
+			PocoData pdThis = PocoData.ForType(typeThis);
+			PocoData pdNext = PocoData.ForType(typeNext);
+
+			// Find split point
+			int firstColumn = pos;
+			var usedColumns = new Dictionary<string, bool>();
+			for (; pos < r.FieldCount; pos++)
+			{
+				// Split if field name has already been used, or if the field doesn't exist in current poco but does in the next
+				string fieldName = r.GetName(pos);
+				if (usedColumns.ContainsKey(fieldName) || (!pdThis.Columns.ContainsKey(fieldName) && pdNext.Columns.ContainsKey(fieldName)))
+				{
+					return pdThis.GetFactory(sql, ConnectionString, firstColumn, pos - firstColumn, r);
+				}
+				usedColumns.Add(fieldName, true);
+			}
+
+			throw new InvalidOperationException(string.Format("Couldn't find split point between {0} and {1}", typeThis, typeNext));
+		}
+
+		// Create a multi-poco factory
+		static Func<IDataReader, object, TRet> CreateMultiPocoFactory<TRet>(Type[] types, string ConnectionString, string sql, IDataReader r)
+		{
+			var m = new DynamicMethod("petapoco_multipoco_factory", typeof(TRet), new Type[] { typeof(MultiPocoFactory), typeof(IDataReader), typeof(object) }, typeof(MultiPocoFactory));
+			var il = m.GetILGenerator();
+
+			// Load the callback
+			il.Emit(OpCodes.Ldarg_2);
+
+			// Call each delegate
+			var dels = new List<Delegate>();
+			int pos = 0;
+			for (int i = 0; i < types.Length; i++)
+			{
+				// Add to list of delegates to call
+				var del = FindSplitPoint(types[i], i + 1 < types.Length ? types[i + 1] : null, ConnectionString, sql, r, ref pos);
+				dels.Add(del);
+
+				// Get the delegate
+				il.Emit(OpCodes.Ldarg_0);													// callback,this
+				il.Emit(OpCodes.Ldc_I4, i);													// callback,this,Index
+				il.Emit(OpCodes.Callvirt, typeof(MultiPocoFactory).GetMethod("GetItem"));	// callback,Delegate
+				il.Emit(OpCodes.Ldarg_1);													// callback,delegate, datareader
+
+				// Call Invoke
+				var tDelInvoke = del.GetType().GetMethod("Invoke");
+				il.Emit(OpCodes.Callvirt, tDelInvoke);										// Poco left on stack
+			}
+
+			// By now we should have the callback and the N pocos all on the stack.  Call the callback and we're done
+			il.Emit(OpCodes.Callvirt, Expression.GetFuncType(types.Concat(new Type[] { typeof(TRet) }).ToArray()).GetMethod("Invoke"));
+			il.Emit(OpCodes.Ret);
+
+			// Finish up
+			return (Func<IDataReader, object, TRet>)m.CreateDelegate(typeof(Func<IDataReader, object, TRet>), new MultiPocoFactory() { _delegates = dels });
+		}
+
+		// Various cached stuff
+		static Cache<string, object> MultiPocoFactories = new Cache<string, object>();
+		static Cache<string, object> AutoMappers = new Cache<string, object>();
+
+		internal static void FlushCaches()
+		{
+			MultiPocoFactories.Flush();
+			AutoMappers.Flush();
+		}
+
+		// Get (or create) the multi-poco factory for a query
+		public static Func<IDataReader, object, TRet> GetFactory<TRet>(Type[] types, string ConnectionString, string sql, IDataReader r)
+		{
+			// Build a key string  (this is crap, should address this at some point)
+			var kb = new StringBuilder();
+			kb.Append(typeof(TRet).ToString());
+			kb.Append(":");
+			foreach (var t in types)
+			{
+				kb.Append(":");
+				kb.Append(t.ToString());
+			}
+			kb.Append(":"); kb.Append(ConnectionString);
+			kb.Append(":"); kb.Append(sql);
+			string key = kb.ToString();
+
+			return (Func<IDataReader, object, TRet>)MultiPocoFactories.Get(key, () =>
+				{
+					return CreateMultiPocoFactory<TRet>(types, ConnectionString, sql, r);
+				}
+			);
+		}
+
+	}
+
+}
