+++ PetaPoco/Core/Sql.cs
@@ -1,0 +1,161 @@
+﻿// PetaPoco - A Tiny ORMish thing for your POCO's.
+// Copyright © 2011-2012 Topten Software.  All Rights Reserved.
+ 
+using System;
+using System.Text;
+using System.Linq;
+using System.Collections.Generic;
+
+namespace PetaPoco
+{
+	// Simple helper class for building SQL statments
+	public class Sql
+	{
+		public Sql()
+		{
+		}
+
+		public Sql(string sql, params object[] args)
+		{
+			_sql = sql;
+			_args = args;
+		}
+
+		public static Sql Builder
+		{
+			get { return new Sql(); }
+		}
+
+		string _sql;
+		object[] _args;
+		Sql _rhs;
+		string _sqlFinal;
+		object[] _argsFinal;
+
+		private void Build()
+		{
+			// already built?
+			if (_sqlFinal != null)
+				return;
+
+			// Build it
+			var sb = new StringBuilder();
+			var args = new List<object>();
+			Build(sb, args, null);
+			_sqlFinal = sb.ToString();
+			_argsFinal = args.ToArray();
+		}
+
+		public string SQL
+		{
+			get
+			{
+				Build();
+				return _sqlFinal;
+			}
+		}
+
+		public object[] Arguments
+		{
+			get
+			{
+				Build();
+				return _argsFinal;
+			}
+		}
+
+		public Sql Append(Sql sql)
+		{
+			if (_rhs != null)
+				_rhs.Append(sql);
+			else
+				_rhs = sql;
+
+			return this;
+		}
+
+		public Sql Append(string sql, params object[] args)
+		{
+			return Append(new Sql(sql, args));
+		}
+
+		static bool Is(Sql sql, string sqltype)
+		{
+			return sql != null && sql._sql != null && sql._sql.StartsWith(sqltype, StringComparison.InvariantCultureIgnoreCase);
+		}
+
+		private void Build(StringBuilder sb, List<object> args, Sql lhs)
+		{
+			if (!String.IsNullOrEmpty(_sql))
+			{
+				// Add SQL to the string
+				if (sb.Length > 0)
+				{
+					sb.Append("\n");
+				}
+
+				var sql = ParametersHelper.ProcessParams(_sql, _args, args);
+
+				if (Is(lhs, "WHERE ") && Is(this, "WHERE "))
+					sql = "AND " + sql.Substring(6);
+				if (Is(lhs, "ORDER BY ") && Is(this, "ORDER BY "))
+					sql = ", " + sql.Substring(9);
+
+				sb.Append(sql);
+			}
+
+			// Now do rhs
+			if (_rhs != null)
+				_rhs.Build(sb, args, this);
+		}
+
+		public Sql Where(string sql, params object[] args)
+		{
+			return Append(new Sql("WHERE (" + sql + ")", args));
+		}
+
+		public Sql OrderBy(params object[] columns)
+		{
+			return Append(new Sql("ORDER BY " + String.Join(", ", (from x in columns select x.ToString()).ToArray())));
+		}
+
+		public Sql Select(params object[] columns)
+		{
+			return Append(new Sql("SELECT " + String.Join(", ", (from x in columns select x.ToString()).ToArray())));
+		}
+
+		public Sql From(params object[] tables)
+		{
+			return Append(new Sql("FROM " + String.Join(", ", (from x in tables select x.ToString()).ToArray())));
+		}
+
+		public Sql GroupBy(params object[] columns)
+		{
+			return Append(new Sql("GROUP BY " + String.Join(", ", (from x in columns select x.ToString()).ToArray())));
+		}
+
+		private SqlJoinClause Join(string JoinType, string table)
+		{
+			return new SqlJoinClause(Append(new Sql(JoinType + table)));
+		}
+
+		public SqlJoinClause InnerJoin(string table) { return Join("INNER JOIN ", table); }
+		public SqlJoinClause LeftJoin(string table) { return Join("LEFT JOIN ", table); }
+
+		public class SqlJoinClause
+		{
+			private readonly Sql _sql;
+
+			public SqlJoinClause(Sql sql)
+			{
+				_sql = sql;
+			}
+
+			public Sql On(string onClause, params object[] args)
+			{
+				return _sql.Append("ON " + onClause, args);
+			}
+		}
+	}
+
+}
