+++ PetaPoco/Core/Transaction.cs
@@ -1,0 +1,33 @@
+﻿// PetaPoco - A Tiny ORMish thing for your POCO's.
+// Copyright © 2011-2012 Topten Software.  All Rights Reserved.
+ 
+using System;
+
+namespace PetaPoco
+{
+	/// <summary>
+	/// Transaction object helps maintain transaction depth counts
+	/// </summary>
+	public class Transaction : IDisposable
+	{
+		public Transaction(Database db)
+		{
+			_db = db;
+			_db.BeginTransaction();
+		}
+
+		public virtual void Complete()
+		{
+			_db.CompleteTransaction();
+			_db = null;
+		}
+
+		public void Dispose()
+		{
+			if (_db != null)
+				_db.AbortTransaction();
+		}
+
+		Database _db;
+	}
+}
