+++ PetaPoco/DatabaseTypes/SQLiteDatabaseType.cs
@@ -1,0 +1,25 @@
+﻿// PetaPoco - A Tiny ORMish thing for your POCO's.
+// Copyright © 2011-2012 Topten Software.  All Rights Reserved.
+
+using System;
+
+
+namespace PetaPoco
+{
+	class SQLiteDatabaseType : DatabaseType
+	{
+		public override object ExecuteInsert(Database db, System.Data.IDbCommand cmd, string PrimaryKeyName)
+		{
+			if (PrimaryKeyName != null)
+			{
+				cmd.CommandText += ";\nSELECT last_insert_rowid();";
+				return db.ExecuteScalarHelper(cmd);
+			}
+			else
+			{
+				db.ExecuteNonQueryHelper(cmd);
+				return -1;
+			}
+		}
+	}
+}
