+++ PetaPoco/DatabaseTypes/SqlServerDatabaseType.cs
@@ -1,0 +1,36 @@
+﻿// PetaPoco - A Tiny ORMish thing for your POCO's.
+// Copyright © 2011-2012 Topten Software.  All Rights Reserved.
+
+using System;
+using System.Linq;
+using System.Text.RegularExpressions;
+
+
+namespace PetaPoco
+{
+	class SqlServerDatabaseType : DatabaseType
+	{
+		public override string BuildPageQuery(long skip, long take, PagingHelper.SQLParts parts, ref object[] args)
+		{
+			parts.sqlSelectRemoved = PagingHelper.rxOrderBy.Replace(parts.sqlSelectRemoved, "");
+			if (PagingHelper.rxDistinct.IsMatch(parts.sqlSelectRemoved))
+			{
+				parts.sqlSelectRemoved = "peta_inner.* FROM (SELECT " + parts.sqlSelectRemoved + ") peta_inner";
+			}
+			var sqlPage = string.Format("SELECT * FROM (SELECT ROW_NUMBER() OVER ({0}) peta_rn, {1}) peta_paged WHERE peta_rn>@{2} AND peta_rn<=@{3}",
+									parts.sqlOrderBy == null ? "ORDER BY (SELECT NULL)" : parts.sqlOrderBy, parts.sqlSelectRemoved, args.Length, args.Length + 1);
+			args = args.Concat(new object[] { skip, skip + take }).ToArray();
+
+			return sqlPage;
+		}
+
+		public override object ExecuteInsert(Database db, System.Data.IDbCommand cmd, string PrimaryKeyName)
+		{
+			cmd.CommandText += ";\nSELECT SCOPE_IDENTITY() AS NewID;";
+			return db.ExecuteScalarHelper(cmd);
+		}
+
+
+	}
+
+}
