+++ PetaPoco/PetaPoco.csproj
@@ -62,34 +62,58 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Models\Generated\Database.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>Database.tt</DependentUpon>
-    </Compile>
+    <Compile Include="Attributes\ColumnAttribute.cs" />
+    <Compile Include="Core\AnsiString.cs" />
+    <Compile Include="Core\ColumnInfo.cs" />
+    <Compile Include="Core\Mappers.cs" />
+    <Compile Include="Core\MultiPocoFactory.cs" />
+    <Compile Include="Core\DatabaseType.cs" />
+    <Compile Include="DatabaseTypes\SQLiteDatabaseType.cs" />
+    <Compile Include="DatabaseTypes\OracleDatabaseType.cs" />
+    <Compile Include="DatabaseTypes\PostgreSQLDatabaseType.cs" />
+    <Compile Include="DatabaseTypes\SqlServerDatabaseType.cs" />
+    <Compile Include="DatabaseTypes\SqlServerCEDatabaseType.cs" />
+    <Compile Include="DatabaseTypes\MySqlDatabaseType.cs" />
+    <Compile Include="Helpers\AutoSelectHelper.cs" />
+    <Compile Include="Helpers\PagingHelper.cs" />
+    <Compile Include="Helpers\ParametersHelper.cs" />
+    <Compile Include="Core\ExpandoColumn.cs" />
+    <Compile Include="Core\PocoColumn.cs" />
+    <Compile Include="Core\PocoData.cs" />
+    <Compile Include="Core\IMapper.cs" />
+    <Compile Include="Core\Transaction.cs" />
+    <Compile Include="Core\Sql.cs" />
+    <Compile Include="Core\TableInfo.cs" />
+    <Compile Include="Core\Page.cs" />
+    <Compile Include="Attributes\PrimaryKeyAttribute.cs" />
+    <Compile Include="Attributes\TableNameAttribute.cs" />
+    <Compile Include="Attributes\ResultColumnAttribute.cs" />
+    <Compile Include="Attributes\IgnoreAttribute.cs" />
+    <Compile Include="Attributes\ExplicitColumnsAttribute.cs" />
     <Compile Include="OracleProvider.cs" />
-    <Compile Include="PetaPoco.cs" />
+    <Compile Include="Database.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Utilities\Cache.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="App.config" />
-    <None Include="PetaPoco.Core.nuspec">
+    <None Include="NuGet\PetaPoco.Core.nuspec">
       <SubType>Designer</SubType>
     </None>
-    <None Include="Models\Generated\PetaPoco.Core.ttinclude" />
-    <None Include="Models\Generated\PetaPoco.Generator.ttinclude" />
-    <None Include="Models\Generated\Database.tt">
-      <Generator>TextTemplatingFileGenerator</Generator>
+    <None Include="T4 Templates\PetaPoco.Core.ttinclude" />
+    <None Include="T4 Templates\PetaPoco.Generator.ttinclude" />
+    <None Include="T4 Templates\Database.tt">
       <LastGenOutput>Database.cs</LastGenOutput>
     </None>
-    <None Include="PetaPoco.nuspec">
+    <None Include="NuGet\PetaPoco.nuspec">
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
   <ItemGroup>
     <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Content Include="TextFile1.txt" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>if $(ConfigurationName) == Release NuGet.exe pack "$(ProjectDir)PetaPoco.nuspec" -o "$(SolutionDir)Output"
