+++ PetaPoco/T4 Templates/Database.tt
