+++ PetaPoco/T4 Templates/PetaPoco.Core.ttinclude
